DefaultOMEROService: tweak identity logic
When the convert method is given an object whose type is actually
compatible with the requested type, we were short-circuiting, doing an
immediate cast. But in the interest of future expansion, we may not
always want to do that. The code later which invokes ClassUtils#convert
will also handle this case just fine, so we can rely on that for now.