Add SCIFIO format for OMERO-backed images
This is similar to Bio-Formats 4's OmeroReader (and OmeroWriter, which
was never finished), for storing and retrieving pixels and metadata to
and from an OMERO database.

It is implemented as a SCIFIO format, making it compatible with ImageJ2.
More specifically, a SCIFIOCellImg can be used to wrap the SCIFIO-backed
data, providing the data as an ImgLib2 Img object, including cached
read-write capabilities on the local machine.

With this format implementation, SCIFIO's ImgOpener class can be used to
"open" (i.e., download on demand) an ImgLib2 ImgPlus directly from an
OMERO server. The ImgPlus will be a SCIFIOImgPlus, which is backed by a
SCIFIOCellImg, which is backed by an OMEROFormat.Reader, which is backed
by an omero.client connection.

The ImgPlus can then be wrapped as an ImageJ2 Dataset, enabling ImageJ2
commands to operate upon it directly.

When changes are made to the local ImgPlus's pixels, those changes
happen in memory, to the Img's "cells"; i.e., paged blocks. As new cells
are requested which push memory consumption beyond desired limits, old
dirty cells are cached out to disk. These cached cells, when present,
are used in preference to data from the original source. In this way, it
is possible to iterate over a massive remote dataset and apply image
processing filters, with all changes recorded to the disk cache, as long
as there is sufficient disk space.

Finally, once processing is complete, SCIFIO's ImgSaver class can be
used to "save" (i.e., upload) the ImgPlus back to OMERO as a new pixels
object.

This first implementation handles only the Reader side of the equation.
The Writer side will be needed to conveniently write ImgLib2 data back
to OMERO using SCIFIO's ImgSaver.