Add a helper class for managing OMERO sessions
This class encapsulates session-specific state, so that the
OMEROFormat.Metadata does not have to do so.

I.e., it is wrong for the Metadata to have sensitive transient state.
Rather, it is the job of the Parser to initialize a session based on the
credentials, then populate the metadata with format-specific information
obtained from the session at parse time. Then the Metadata's
populateImagMetadata can translate that format-specific information into
ImageMetadata without needing access to the OMERO session.