Remove clever logic to glean command from filename
Unfortunately, OMERO calls each script as "./script" so the trick of
using argv won't work.

Of course, we could change OMERO to call it as the original filename,
but then we'd have to worry about edge cases with spaces, behavior on
Windows, etc. So for now we hardcode the commands again.