Always use image IDs, not pixels IDs
The language and code was rather inconsistent before. Modern OMERO heavily
prefers use of image IDs, since each Image can now only have one Pixels.
They are the same 99% of the time but it is good to take extra care that
we always want an image ID, not a pixels ID, in the API.