ScriptGenerator: skip "non-app" modules
If a module does not have the "application" menu root, we skip it.
Otherwise, context menu items and all sorts of other cruft will
potentially be exposed by the generator. We only want "real" ImageJ
commands that are in the menus, right now.

This change also refactors the validity checking into its own method.