ScriptGenerator: use base OMERO dir as arg
Previously, the script generator accepted the exact path (e.g.,
$OMERO_DIR/lib/scripts/imagej where OMERO script stubs were supposed
to go). But it makes more sense to accept the root directory of the
OMERO installation, verify that the argument is in fact valid, and
then generate the scripts in the appropriate subdirectory always.

With this change, the generator now exits with a non-zero error code
if something goes wrong.