Remove obsolete launcher scripts
We definitely do not want to be wrapping the "jython" executable with
our own hacky shell script. And the "imagej" launcher is no longer
necessary now that we are distributing all required OMERO JARs with the
OMERO-5.0 update site, and no longer need to add JARs from the actual
OMERO installation's lib/server directory to the classpath.