ScriptGenerator: make headlessOnly a field
It was previously passed to the generateAll method as a functional
parameter. But each ScriptGenerator instance is already stateful (it has
an ImageJ application context), so let's just make the headlessOnly flag
part of that state. This paves the way for adding additional pieces of
state without polluting the generateAll method signature any further.