OMEROFormat: prioritize the OMERO format
This is a hack to work around the fact that OBFFormat (and maybe others)
throw an exception when given an OMERO source string (i.e., starting
with "omero:" and/or ending with ".omero"). Of course, that bug must be
fixed, but for now, we avoid it by making sure OMEROFormat goes first.