OMEROService: revise how saver metadata is queried
Mark and I settled on a different approach for ImgSaver to provide
access to the Metadata used by the Writer. Rather than writing it into
the original ImgPlus's properties table -- which destroys the original
Metadata put there by the ImgOpener, can suffer from race conditions,
and be misleading with respect to its contents -- we instead have the
ImgSaver#saveImg methods return the Metadata. And ImageJ's
DatasetService#save does the same thing.