OMEROFormat.Metadata: enable Image+Pixels caching
After loading an Image or Pixels descriptor from a remote OMERO server,
it is very nice to have somewhere to cache it. The Metadata is the
natural place, since doing it in the OMEROSession would prevent the
session being used to populate multiple different Metadata objects.

However, we do not want the cached Image or Pixels to be persisted or
otherwise treated as 1st class metadata fields. So they are not
annotated with @Field. They are also nulled out when the metadata's
respective imageID or pixelsID is updated.

If consuming code needs to be certain it will receive a non-null Image
and/or Pixels, it should use the OMEROSession's respective loadImage
and/or loadPixels methods, which load those objects on demand.