WIP: Set the global channel mins and maxes
This lets OMERO's initial rendering settings match what you were
looking at in ImageJ at the time of upload.

Right now, this is just a placeholder to remind that doing this would be
nice. There are some architectural considerations though, with respect
to SCIFIO's ImgSaver working with ImgPlus while ImageJ2's channel
scalings are part of DatasetView (not Dataset).