Eliminate usages of net.imagej.ImageJ
This change is in preparation for reducing the dependencies of the
imagej-omero component. While net.imagej:imagej is very convenient as a
starting point and for casual developers, for robust library components
like imagej-omero it makes more sense to do the dependencies properly,
declaring only those dependencies which are actually needed.