README.md: use OMERO-5.1, not OMERO-5.0
We do leave a mention of the old OMERO-5.0 update site, although
further updates are not currently planned to that update site.

In the future, if and when the OMERO API continues to evolve and break
backwards compatibility again, we'll have to consider whether to start
maintaining multiple integration branches for ImageJ-OMERO. But not yet.