Modify code to attach new images to datasets
This helps deal with the orphan image issues. Now when an ImageJ script
creates a new image, it will be attached to the same dataset(s) as the
input image(s). Additionally if a valid dataset is provided when
exporting an image from ImageJ to OMERO, that image will now be attached
to that OMERO dataset. As part of this change, methods for getting a
gateway and experimenter were added to OMEROSession.

OMEROSession: avoid use of third party Log class