Set column header in populateImageJColumn method
Since createImageJColumn is no longer called for all instances of
creating an imageJ column, the headers were not being set. This change
makes it so the populateImageJColumn method checks if the column has a
header and if it doesn't sets the header to be the name of the OMERO
column.