Modify upload/download table to use SecurityContext
This causes a new error to be thrown. As such, numerous files had to be
changed. In addition to that, this changes the order of calls in those
methods which breaks the mocking in the tests. Hence, this commit also
fixes the tests. Finally, mocking a Gateway requires
com.google.collections to be on the class path.