Modify OMERORefColumn to store original data array
When the TablesFacility gets a table, it takes the reference type
columns and makes DataObjects for each reference. Then, when this table
is put back on the OMERO server these refence objects once again store
the ids of the objects. So in imagej-omero the reference columns will
still be LongColumns (at least for now) but will also keep the array of
the DataObjects. This is mostly done to simplify uploading these columns
back to the OMERO server.