Speed up build by caching .m2
This makes use of the (newly introduced) DOCKER_ARGS as an
extension point in omero-test-infra to allow mounting
arbitrary locations. The reason for the environment variable
is that the repo's Dockerfile could be using an arbitrary
user. (Another strategy would be to use a well-known
location like /m2 and then set the mvn property.)

This additionally *doesn't* run the mvn install during
`docker build` since it's not possible to mount the m2
leading to a double download of all the dependencies.