Remove OMERORoi and corresponding converter
This structure treats all OMERO ROIData objects as union composites.
However, this isn't really accurate. Instead, we should treat the
ROIData as a collection of ShapeData. Additionally, this setup didn't
solve the 2D vs. 5D problem. All ShapeData should be converted to 2D
MaskPredicates which can be projected into 5D space. And lastly, this
structure was a bit confusing and the logic was difficult to follow.