SaveToOMERO: name image parameter "image"
The name "dataset" is ambiguous/confusing between ImageJ and OMERO,
because in OMERO a "dataset" is a collection of N-dimensional images,
whereas in ImageJ a "dataset" is (currently) one N-dimensional image.

The OpenFromOMERO command calls the net.imagej.Dataset parameter
"image" and we can do the same here, to avoid confusion.

This change breaks existing macros calling File > Export > OMERO...
but now (rather than later) is the time to do that.

Closes #100.