Workaround SSLHandshakeException with OMERO 5.4
OMERO uses Ice's ADH settings to connect without the need
to install a certificate by default. Newer JDKs are disabling
more and more ciphers. In order to continue working with older
server versions, the ciphers need to be added back *before*
any other code loads the SSL subsystem in a JVM. This patch
places a static block at the beginning of all classes which
call `new omero.client` in order to be as defensive as possible.

see:
 - https://travis-ci.org/imagej/imagej-omero/jobs/529909300#L3040
 - https://github.com/openmicroscopy/openmicroscopy/pull/5949