Update to the new SCIFIO API
This updates the parent POM to pom-scijava 29.0.0-beta-2.

We don't have String ids any more, but rather Locations and DataHandles.

This commit also removes OMEROSession from the JMockit Expectations.
Having instantiation without assignment appears to be non-kosher in the
latest JMockit. Removing this invocation did not cause tests to fail, so
it's not clear why it was originally needed.

See also:
* https://stackoverflow.com/q/16576177
* https://gitmemory.com/issue/jmockit/jmockit1/630/531445943
* https://gitmemory.com/issue/jmockit/jmockit1/630/531494884

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>